% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IFPCA.R
\name{ifpca}
\alias{ifpca}
\title{Intensity Functional Principal Component Analysis (IFPCA)}
\usage{
ifpca(time, fu_train, fu_valid, PPIC_K = FALSE, n.grid = 401, propvar = 0.85)
}
\arguments{
\item{time}{longitudinal encounter times}

\item{fu_train}{follow-up time (training)}

\item{fu_valid}{follow-up time (validation)}

\item{PPIC_K}{a logical indicating whether you want to use Pseudo-Poisson Information Criterion to choose 
the number of principal components K (K.select="PPIC") \code{TRUE} or another criterion to choose 
K (K.select="PropVar") \code{FALSE} in the PP_FPCA_Parallel function (hidden). Default is \code{FALSE}.}

\item{n.grid}{an integer value for grid points used in estimating covariance function g. Default is \code{401}.}

\item{propvar}{a proportion of variation used to select number of FPCs. Default is \code{0.85}.}
}
\description{
Performs IFPCA to extract features from longitudinal encounter data.
}
